Check permission handling.

+++ classes/content_user_data.php
@@ -39,6 +39,8 @@ class content_user_data {
      * @throws \coding_exception
      */
     public static function handle_ajax() {
+        global $DB;
+
         // Query String Parameters.
         $content_id = required_param('content_id', PARAM_INT);
         $data_id = required_param('data_type', PARAM_ALPHA);
@@ -60,10 +62,28 @@ class content_user_data {
             exit;
         }
 
+        $context = \context_course::instance($DB->get_field('hvp', 'course', array('id' => $content_id)));
+
         // Delete user data.
         if ($data === '0') {
+
+            // Check permissions
+            if (!has_capability('mod/hvp:deletecontentuserdata', $context)) {
+                \H5PCore::ajaxError(get_string('nopermissiontodeletecontentuserdata', 'hvp'));
+                http_response_code(403);
+                exit;
+            }
+
             self::delete_user_data($content_id, $sub_content_id, $data_id);
         } else {
+
+            // Check permissions
+            if (!has_capability('mod/hvp:savecontentuserdata', $context)) {
+                \H5PCore::ajaxError(get_string('nopermissiontosavecontentuserdata', 'hvp'));
+                http_response_code(403);
+                exit;
+            }
+
             // Save user data.
             self::save_user_data($content_id, $sub_content_id, $data_id, $pre_load, $invalidate, $data);
         }
